use supported columns by importer for template
In fact, this is one of the most important changes in this PR.
In this way we achieve full compatibility and a single source of input data,
i.e. it is a kind of interface high level definition for users. In practice,
this is expressed by the fact that it will not be possible to add a new column
to the template without creating a mapper for it. That's it!

This removed the corresponding column list file and was another step
in reducing the responsibility of XLS services, now it is
delegated to the exporter and importer.